Big and incomplete file
lots of writing about stuff planned for the program in the distant
distant future relating to how character scripting works.
It is messy and also not finished, and will soon either become split up
into multiple files, or just better organised with sub headings.
At the moment it serves me more than anyone else trying to make sense of
it.