OOP fixes #5 and descriptions fixes #7
spaces are now 3 classes for the sake of SRP, which seems sufficient for now, and definitely encourages the planned changes I couldn't be stuffed doing

descriptions are now much nicer as well, but only two of the descriptions are possible run-time, out of eight possible descriptions. This will change as more actions are implemented, and the existing descriptions become relevant.

various stubs for graspers are also left in this commit, and their unmotivated stubness is the reason why I needed to do some SRP on spaces. (messy classes are messy)