IO restructuring!
IO is now done through special classes to which the player points, and input is now a weird precarious feedback loop that could end if not tended to properly.

In addition, Events now use Spaces to determine their info instead of the Space passing itself into an Event it creates.