Grabbing!
(and touching which is a generalization of grabbing)
It is now possible to pick up items, putting them in your hand. When you pick up an item, whatever was in your hand is dropped.

Currently when you pick up and drop items you are double informed, since sight is separate to touch; this is part of the charm, and a solution will be the goal of an advanced UI